ast::Text optimization
most Text instances will only be one fragment. More fragments are only
needed when there are comments or scaped characters to skip. Using a
smallvec with space for 1 fragment in the stack reduces the allocations
needed.

In the benchmark recipe, in my computer, this was a small
improvement and to be honest maybe not even worthit. But doing it it
brings peace into my mind.