Add fuzzing for the rust parser
This adds a fuzzing task for the rust parser.  This is run as release
to avoid various debug assertions in the project.  I've had this
running on my workstation and it hasn't found anything, but it is
useful to keep in CI to catch regressions.

I recongize that fuzzing adds extra time to the pipeline, so I did a
few things to try and minimize the impact.

* First, it is a seperate, parallel workflow.  Since the main build
  and test is a debug build and fuzz is a release build there is very
  little wasted work in rebuilding.
* Second, I added caches for both tasks.  This should help keep build
  time town in general.
* The time the fuzzing task will run is capped at 2 minutes.  This is
  pretty short.  This was picked to not be too much longer than
  previous runs.  This is a bit of a trade off.  The longer this
  value is, the better the coverage.

Later I will take a look at having it excercise more parser
configuration options to get better coverage.  At first I just wanted
to try and cover the default case well.